+++ libyara/modules/elf.c
@@ -205,7 +205,7 @@ uint64_t elf_rva_to_offset_##bits##_##bo(                                      \
 #define PARSE_ELF_HEADER(bits,bo)                                              \
 void parse_elf_header_##bits##_##bo(                                           \
   elf##bits##_header_t* elf,                                                   \
-  size_t base_address,                                                         \
+  uint64_t base_address,                                                       \
   size_t elf_size,                                                             \
   int flags,                                                                   \
   YR_OBJECT* elf_obj)                                                          \
